Move default button implementation from TriangleButton to OsuButton
